功能拆分到了key-value类，简化了blob类和tree类
